+++ CHANGELOG.md
@@ -4,11 +4,20 @@ All notable changes to the "tyranosyntax" extension will be documented in this f
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.14.0] 2025-09-27
+
+- 設定に`TyranoScript syntax.tyranoBuilder.enabled`を追加しました。[issue#336](https://github.com/orukRed/tyranosyntax/issues/336)
+  - ティラノビルダーを使用している場合はtrueに設定することで、ティラノビルダーに存在する固有のパラメータ（_clickable_imgなど）でエラーが出なくなります。
+- たまにmacroタグで定義したタグを認識してくれない問題への暫定的対応を行いました。[issue#334](https://github.com/orukRed/tyranosyntax/issues/334)
+  - ただし、再現性が低く確認ができていないことから解決していない可能性があります。
+  - 同様のエラーが発生する場合は再度ご報告いただけますと嬉しいです。
+
 ## [1.13.0] 2025-09-15
 
 - ティラノビルダー固有のタグとパラメータをエラーとして検出しないようにしました。[issue#326](https://github.com/orukRed/tyranosyntax/issues/326)
   - ティラノビルダー専用タグ（`tb_eval`、`tb_show_message_window`など）がエラーとして検出されないようになりました。
-  - ティラノビルダー専用パラメータ（`playbgm`タグの`time`、`fadein`パラメータなど）がエラーとして検出されないようになりました。  - 設定で`TyranoScript syntax.tyranoBuilder.skipTags`および`TyranoScript syntax.tyranoBuilder.skipParameters`をカスタマイズ可能です。
+  - ティラノビルダー専用パラメータ（`playbgm`タグの`time`、`fadein`パラメータなど）がエラーとして検出されないようになりました。設定で`TyranoScript syntax.tyranoBuilder.skipTags`および`TyranoScript syntax.tyranoBuilder.skipParameters`をカスタマイズ可能です。
+
 
 ## [1.12.1] 2025-09-08